requirements.txt
aws-cdk.core==1.133.0
aws-cdk.assertions==1.133.0
aws_cdk.aws_dynamodb==1.133.0
aws_cdk.aws_apigateway==1.133.0
aws_cdk.aws_iam==1.133.0
aws_cdk.aws_cognito==1.133.0